### 2025-04-03 #7976 @fruzsinaagocs
👋 @Sosnowsky, thank you for your submission!

I'd like to ask you to do a few things before we move to the review phase:
- If you have any reviewers in mind who could look at your submission, please name them below without @-mentioning them. A list of JOSS reviewers can be found [here](https://reviewers.joss.theoj.org/reviewers).
- Please check the missing/invalid DOI-s flagged by the editorialbot above.
- Please also correct this typo in the reference list: `lower_power_1990` should read `lowen_power_1990`.
- Once you have corrected the DOIs and the citation, please generate a pdf again by commenting `@editorialbot generate pdf`.
