### 2025-04-04 2dd8965
Fix bib sources
