+++ paper/paper.bib
@@ -7,6 +7,7 @@
     pages = {095102},
     year = {2024},
     month = {09},
+    doi = {10.1063/5.0197251},
 }
 
 @article{dippolito_convective_2011,
@@ -16,34 +17,33 @@
 	journal = {Physics of Plasmas},
 	author = {D'Ippolito, D. A. and Myra, J. R. and Zweben, S. J.},
 	year = {2011},
+	doi = {10.1063/1.3594609},
 }
 
 @article{fiedler_coherent_1988,
-title = {Coherent structures in turbulent flows},
-journal = {Progress in Aerospace Sciences},
-volume = {25},
-number = {3},
-pages = {231-269},
-year = {1988},
-issn = {0376-0421},
-doi = {https://doi.org/10.1016/0376-0421(88)90001-2},
-url = {https://www.sciencedirect.com/science/article/pii/0376042188900012},
-author = {H.E. Fiedler},
+    title = {Coherent structures in turbulent flows},
+    journal = {Progress in Aerospace Sciences},
+    volume = {25},
+    number = {3},
+    pages = {231-269},
+    year = {1988},
+    issn = {0376-0421},
+    doi = {10.1016/0376-0421(88)90001-2},
+    author = {H.E. Fiedler},
 }
 
 @inproceedings{nosov_coherent_2009,
-author = {V. V. Nosov and V. M. Grigoriev and P. G. Kovadlo and V. P. Lukin and E. V. Nosov and A. V. Torgaev},
-title = {{Coherent structures in turbulent atmosphere}},
-volume = {7296},
-booktitle = {Fifteenth International Symposium on Atmospheric and Ocean Optics/Atmospheric Physics},
-editor = {Gennadii G. Matvienko and Yurii N. Ponomarev},
-organization = {International Society for Optics and Photonics},
-publisher = {SPIE},
-pages = {729609},
-keywords = {turbulence, spectrum, coherent structure, Benard cell, stochastization scenario},
-year = {2009},
-doi = {10.1117/12.823804},
-URL = {https://doi.org/10.1117/12.823804}
+    author = {V. V. Nosov and V. M. Grigoriev and P. G. Kovadlo and V. P. Lukin and E. V. Nosov and A. V. Torgaev},
+    title = {{Coherent structures in turbulent atmosphere}},
+    volume = {7296},
+    booktitle = {Fifteenth International Symposium on Atmospheric and Ocean Optics/Atmospheric Physics},
+    editor = {Gennadii G. Matvienko and Yurii N. Ponomarev},
+    organization = {International Society for Optics and Photonics},
+    publisher = {SPIE},
+    pages = {729609},
+    keywords = {turbulence, spectrum, coherent structure, Benard cell, stochastization scenario},
+    year = {2009},
+    doi = {10.1117/12.823804},
 }
 
 @article{zweben_invited_2017,
@@ -56,8 +56,7 @@ URL = {https://doi.org/10.1117/12.823804}
 	author = {Zweben, S. J. and Terry, J. L. and Stotler, D. P. and Maqueda, R. J.},
 	year = {2017},
 	pmid = {28456269},
-        pages = {041101},
-	file = {Attachment:/home/sosno/Zotero/storage/VGNCMUZA/zweben_rsi_88_041101.pdf:application/pdf},
+    pages = {041101},
 }
 
 @article{offeddu_analysis_2023,
@@ -69,11 +68,8 @@ URL = {https://doi.org/10.1117/12.823804}
     pages = {033512},
     year = {2023},
     month = {03},
[…]
