### 2025-04-04 #7976 @Sosnowsky
@fruzsinaagocs Thanks for your comments! I fixed the references and generated a new pdf.
I don't have any reviewers in mind, but can find some under the keyboards of plasma physics, or complex systems. For example: EmilyBourne, michaeltouati, EwoutH, SunnyXu.
