### 2025-04-09 #8032 @editorialbot
[REVIEW]: Blobmodel: A Python package for generating superpositions of pulses in one and two dimensions
**Submitting author:** @Sosnowsky (<a href="http://orcid.org/0000-0003-2054-1384">Juan Losada</a>)
**Repository:** https://github.com/uit-cosmo/blobmodel
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.1.0
