### 2025-04-17 uit-cosmo/blobmodel#104 @EmilyBourne
[JOSS] Problem running the example in "Getting Started"
This issue relates to an ongoing reviews at: https://github.com/openjournals/joss-reviews/issues/8032

I am trying to go through the tutorial here:
https://blobmodel.readthedocs.io/en/latest/getting_started.html

but I am seeing unexpected errors. My code is:
```python
import matplotlib.pyplot as plt
from blobmodel import Model

if __name__ == '__main__':
    bm = Model(
        Nx=100,
        Ny=100,
        Lx=10,
        Ly=10,
        dt=0.1,
        T=20,
        periodic_y=True,
        blob_shape="gauss",
        num_blobs=100,
        t_drain=1e10,
        t_init=10,
    )
    print(bm)
    ds = bm.make_realization(file_name="example.nc")

    ds["n"].isel(y=0).mean(dim=("t")).plot()
    plt.show()
```

The output is:
```
2d Blob Model with num_blobs:100 and t_drain:10000000000.0
Summing up Blobs:   0%|                                                                                                                              | 0/100 [00:00<?, ?it/s]
Traceback (most recent call last):
  File "/home/emily/Code/test/tmp/tmp.py", line 19, in <module>
    ds = bm.make_realization(file_name="example.nc")
         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/emily/Code/venv_blobs/lib/python3.11/site-packages/blobmodel/model.py", line 209, in make_realization
    self._sum_up_blobs(blob, speed_up, error)
  File "/home/emily/Code/venv_blobs/lib/python3.11/site-packages/blobmodel/model.py", line 274, in _sum_up_blobs
    _single_blob = blob.discretize_blob(
                   ^^^^^^^^^^^^^^^^^^^^^
  File "/home/emily/Code/venv_blobs/lib/python3.11/site-packages/blobmodel/blobs.py", line 162, in discretize_blob
    self._single_blob(
  File "/home/emily/Code/venv_blobs/lib/python3.11/site-packages/blobmodel/blobs.py", line 240, in _single_blob
    * self._propagation_direction_shape(
      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/emily/Code/venv_blobs/lib/python3.11/site-packages/blobmodel/blobs.py", line 318, in _propagation_direction_shape
    return self.blob_shape.get_blob_shape_prop(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'str' object has no attribute 'get_blob_shape_prop'
```

It seems that the suggested line `blob_shape="gauss",` doesn't match up with the current implementation. Maybe there is a step missing in the tutorial using the `blob_shape` module?

The following code seems to work better:
```python
import matplotlib.pyplot as plt
from blobmodel import Model, BlobShapeImpl, BlobShapeEnum

if __name__ == '__main__':
    bm = Model(
        Nx=100,
        Ny=100,
        Lx=10,
        Ly=10,
        dt=0.1,
        T=20,
        periodic_y=True,
        blob_shape=BlobShapeImpl(BlobShapeEnum.gaussian),
        num_blobs=100,
        t_drain=1e10,
        t_init=10,
    )
    print(bm)
    ds = bm.make_realization(file_name="example.nc")

    ds["n"].isel(y=0).mean(dim=("t")).plot()
    plt.show()
```
