### 2025-04-17 gregordecristoforo/2d_propagating_blobs#105 @EmilyBourne
[JOSS] Documentation issues
This issue relates to an ongoing reviews at: https://github.com/openjournals/joss-reviews/issues/8032

Related to #104. It seems that `Model` does not accept a string as the `blob_shapes` parameter.

This page of the documentation therefore needs updating:
https://blobmodel.readthedocs.io/en/latest/blob_shapes.html
