### 2025-04-17 uit-cosmo/blobmodel#106 @EmilyBourne
[JOSS] Missing documentation
This issue relates to an ongoing reviews at: https://github.com/openjournals/joss-reviews/issues/8032

The function `Model.make_realization` outputs data in a NetCDF format. The docstring for this method states:
> file_name (str, optional) – File name for the .nc file containing data as an xarray dataset
> Returns : xarray dataset with the resulting data.

But I can't seem to find a list of what data is described here. Grid? Distribution function? Density? Something else?
