### 2025-04-17 #8032 @EmilyBourne
@fruzsinaagocs I have completed my first pass of the submission. I have opened 4 issues and left a review on the PR of the paper itself with some grammar suggestions.

Globally the submission seems ok but the documentation seems to be out of date. This needs fixing before proceeding
