### 2025-04-22 19b7de6
Add statement of field https://github.com/uit-cosmo/blobmodel/issues/103
