### 2025-04-22 1c26f0e
Updated docs to new interfaces and created doc tests https://github.com/uit-cosmo/blobmodel/issues/104
