### 2025-04-22 cb57e41
Updated docs to new interfaces and created doc tests https://github.com/uit-cosmo/blobmodel/issues/104
