### 2025-04-22 uit-cosmo/blobmodel#104 @Sosnowsky
I fixed this and all the other docs examples in https://github.com/uit-cosmo/blobmodel/commit/cb57e4109355cfc44b6dfa64b5eb638a659b30ce

Also created a new test suite to test that the doc examples run.
