### 2025-04-22 9fdf61d
Update examples to new interface https://github.com/uit-cosmo/blobmodel/issues/104
