+++ examples/compare_to_analytical_sol.py
@@ -1,11 +1,22 @@
-from blobmodel import Model, DefaultBlobFactory
+from blobmodel import (
+    Model,
+    DefaultBlobFactory,
+    DistributionEnum,
+    BlobShapeImpl,
+    BlobShapeEnum,
+)
 import xarray as xr
 import matplotlib.pyplot as plt
 import numpy as np
 import os
 
 # use DefaultBlobFactory to define distribution functions fo random variables
-bf = DefaultBlobFactory(A_dist="deg", wx_dist="deg", vx_dist="deg", vy_dist="zeros")
+bf = DefaultBlobFactory(
+    A_dist=DistributionEnum.deg,
+    wx_dist=DistributionEnum.deg,
+    vx_dist=DistributionEnum.deg,
+    vy_dist=DistributionEnum.zeros,
+)
 
 tmp = Model(
     Nx=100,
@@ -14,7 +25,7 @@ tmp = Model(
     Ly=0,
     dt=1,
     T=1000,
-    blob_shape="exp",
+    blob_shape=BlobShapeImpl(BlobShapeEnum.exp),
     t_drain=2,
     periodic_y=False,
     num_blobs=10000,
