+++ examples/2_sided_exp_pulse.py
@@ -1,21 +1,22 @@
-from blobmodel import Model, show_model, DefaultBlobFactory, BlobShapeImpl
+from blobmodel import (
+    Model,
+    show_model,
+    DefaultBlobFactory,
+    BlobShapeImpl,
+    DistributionEnum,
+    BlobShapeEnum,
+)
 import matplotlib.pyplot as plt
-import numpy as np
 
 
 bf = DefaultBlobFactory(
-    A_dist="deg",
-    vy_dist="deg",
-    wx_dist="deg",
-    spx_dist="deg",
-    spy_dist="deg",
+    A_dist=DistributionEnum.deg,
+    wx_dist=DistributionEnum.deg,
+    spx_dist=DistributionEnum.deg,
+    spy_dist=DistributionEnum.deg,
     shape_param_x_parameter=0.5,
     shape_param_y_parameter=0.5,
-    blob_alignment=False,
 )
-
-bf.set_theta_setter(lambda: np.random.uniform(-np.pi / 2, np.pi / 2))
-# bf.set_theta_setter(lambda: np.pi/4)
 
 bm = Model(
     Nx=100,
@@ -25,8 +26,7 @@ bm = Model(
     dt=0.1,
     T=10,
     num_blobs=10,
-    blob_shape=BlobShapeImpl("exp", "2-exp"),
-    periodic_y=True,
+    blob_shape=BlobShapeImpl(BlobShapeEnum.double_exp, BlobShapeEnum.double_exp),
     t_drain=1e10,
     blob_factory=bf,
 )
