+++ examples/custom_blobfactory.py
@@ -4,6 +4,8 @@ from blobmodel import (
     Blob,
     show_model,
     AbstractBlobShape,
+    BlobShapeImpl,
+    BlobShapeEnum,
 )
 import numpy as np
 
