### 2025-04-22 uit-cosmo/blobmodel#104 @Sosnowsky
Fixed also the files under examples folder https://github.com/uit-cosmo/blobmodel/commit/9fdf61d1dc3df6990b829bda6e5e620fd5f69d71
