### 2025-04-22 uit-cosmo/blobmodel#104 @EmilyBourne
I've just noticed that you have put these commits on the `joss` branch. I don't think they affect the contents of the paper even if they are relevant for the review. Usually I wait until such changes are available on the main branch (and can therefore be seen in the docs) before ticking of such items in my review. Merging the doc corrections to the main will also be useful for the other reviewer so he doesn't leave similar issues when he reviews
