### 2025-04-22 uit-cosmo/blobmodel#104 @Sosnowsky
That's right! I created now a new branch https://github.com/uit-cosmo/blobmodel/tree/docs_update
I cherry-picked the commits mentioned in this issue there and will add some further changes relevant for this issue (mainly I want to find a solution to now repeat the code in the docs, tests and examples, as you mentioned in the commit)
