### 2025-04-23 407852a
Use enums in blob shapes docs
