### 2025-04-23 0aae5f8
Improve description of xarray return value of the model https://github.com/uit-cosmo/blobmodel/issues/106
