+++ docs/getting_started.rst
@@ -45,7 +45,7 @@ If we provide a ``file_name`` to the ``make_realization`` method, it will store 
 
    ds = bm.make_realization(file_name="example.nc")
 
-The ``make_realization`` mehtod can take two more arguments, ``speed_up`` and ``error``, which can be helpful for integrating very large datasets. 
+The ``make_realization`` method can take two more arguments, ``speed_up`` and ``error``, which can be helpful for integrating very large datasets.
 By setting ``spee_up`` to ``True``, the code will truncate the blobs when the blob values fall under the given ``error`` value. 
 The code assumes an exponential shape for the blobs when calculating the truncation position (see :ref:`blob-shapes` for further details).
 The resulting realization will look something like this:
@@ -60,7 +60,10 @@ Which may be a bit boring, as all the blobs move with the same speed and have th
 
 
 The ``make_realization`` method returns the realization in the form of an `xarray dataset <https://docs.xarray.dev/en/stable/index.html>`_. 
-The superposed pulses are stored in the ``n`` variable of the dataset. We can now analyse the data using the convenient xarray syntax, for example:
+The superposed pulses are stored in the ``n`` variable of the dataset. The dimension coordinates are ``x``, ``y`` and
+``t`` for the horizontal, vertical and time coordinate, respectively.
+
+We can now analyse the data using the convenient xarray syntax, for example:
 
 .. code-block:: python
 
