### 2025-04-23 uit-cosmo/blobmodel#106 @Sosnowsky
Addressed in https://github.com/uit-cosmo/blobmodel/commit/0aae5f8d9740f360f3aec7690d336f2615a0e44a
Same pull request as other documentation issues.
