+++ docs/drainage_time.rst
@@ -8,29 +8,12 @@ We can also set ``t_drain`` to an array like of length ``Nx``. In this case ``t_
 
 Let's take a look at a quick example. Let's assume we want ``t_drain`` to decrease linearly with x. We could implement this as follows:
 
-.. code-block:: python
+.. literalinclude:: ../tests/test_docs.py
+   :language: python
+   :start-after: # PLACEHOLDER drainage_time_0
+   :end-before: # PLACEHOLDER drainage_time_1
 
-  from blobmodel import Model, DefaultBlobFactory
-  import numpy as np
-
-  t_drain = np.linspace(2, 1, 100)
-
-  tmp = Model(
-      Nx=100,
-      Ny=1,
-      Lx=10,
-      Ly=0,
-      dt=1,
-      T=1000,
-      blob_shape=BlobShapeImpl(BlobShapeEnum.exp),
-      t_drain=t_drain,
-      periodic_y=False,
-      num_blobs=100,
-  )
-
-  ds_changing_t_drain = tmp.make_realization()
-
-The time averaged x-profile of ``n`` compared to a constant ``t_drain`` = 2 would then look like this: 
+The time averaged x-profile of ``n`` compared to a constant ``t_drain`` = 2 would then look like this:
 
 .. image:: change_t_drain_plot.png
    :scale: 80%
