### 2025-04-24 uit-cosmo/blobmodel#104 @Sosnowsky
@EmilyBourne I'm done with the updates of the documentation. I am a bit unsure about the procedure now, but is it okay if I ask you to check whether the pull request https://github.com/uit-cosmo/blobmodel/pull/107 addresses this issues (and the other two documentation related issues) before I merge into the main branch?
Thanks!
