### 2025-04-29 uit-cosmo/blobmodel#104 @EmilyBourne
@Sosnowsky I have had a look. I think this adresses my issues (I am more sure about the other two issues). But it is a bit hard to read in the rst file. I am happy for you to merge and close these issues. I can always reopen or create a new issue if I still cannot run the example
