### 2025-04-30 uit-cosmo/blobmodel#105 @Sosnowsky
Fixed on https://github.com/uit-cosmo/blobmodel/pull/107
