### 2025-04-30 uit-cosmo/blobmodel#106 @Sosnowsky
Fixed on https://github.com/uit-cosmo/blobmodel/pull/107
