### 2025-05-15 #8032 @andrewgiuliani
Hi @fruzsinaagocs, I have completed my first pass.  I have opened 5 issues and left comments on their manuscript in the their PR containing the write-up.

Overall, the code appears to work as described and installed without issues on my machine.  My major comments regard improving docstrings for examples and unit tests.  I'm happy to take another look once the authors have had time to address my comments.
