### 2025-05-16 d9eb7f3
Remove if True statement and use rotate function
