### 2025-05-16 e66e29c
Print mypy version
