### 2025-05-16 60932e7
Revert "Print mypy version"
This reverts commit e66e29c453c3255f3585fb38f9e73b98720be397.
