### 2025-05-16 ca3d2e2
Revert "Make mypy happy"
This reverts commit 790bb611d7b6d3920809145218cbca51a5dc2ef4.
