### 2025-05-16 2b15906
Revert "Make mypy happy"
This reverts commit 53a0c31e6883786957ab0244ecdb90b98c2abb4f.
