### 2025-05-16 bb89147
Revert "Remove if True statement and use rotate function"
This reverts commit d9eb7f3544a163b1ffb12d26d7c3505b9a2c89c7.
