### 2025-05-16 673fc09
Remove true statement
