### 2025-05-16 fb9a2f1
Add docstrings to some unit tests, gotta catch a bus
