### 2025-05-20 9a3b536
Added assertions for type checking on costum classes
