+++ tests/test_one_dim.py
@@ -56,21 +56,17 @@ def test_1d_warning():
     Checks model warning when instantiating a one-dimensional model with Ny!=0.
     """
     with pytest.warns(UserWarning):
-        for vy_dist in ["deg", "zeros"]:
-            bf = DefaultBlobFactory(
-                A_dist="deg", wx_dist="deg", vx_dist="deg", vy_dist=vy_dist
-            )
-            one_dim_model = Model(
-                Nx=100,
-                Ny=100,
-                Lx=10,
-                Ly=10,
-                dt=1,
-                T=1000,
-                blob_shape=BlobShapeImpl(BlobShapeEnum.exp, BlobShapeEnum.gaussian),
-                t_drain=2,
-                periodic_y=False,
-                num_blobs=1,
-                blob_factory=bf,
-                one_dimensional=True,
-            )
+        Model(
+            Nx=100,
+            Ny=100,
+            Lx=10,
+            Ly=10,
+            dt=1,
+            T=1000,
+            blob_shape=BlobShapeImpl(BlobShapeEnum.exp, BlobShapeEnum.gaussian),
+            t_drain=2,
+            periodic_y=False,
+            num_blobs=1,
+            blob_factory=bf,
+            one_dimensional=True,
+        )
