### 2025-05-20 1f9f207
Add docstrings to examples
