+++ examples/blob_tilting.py
@@ -8,7 +8,8 @@ from blobmodel import (
 )
 import numpy as np
 
-# To have something to compare we will first make a simulation with aligned blobs moving with velocities
+# Example with tilted blobs, to have something to compare we will first make a simulation with
+# aligned blobs moving with velocities
 
 vx = 1
 vy = 1
