+++ examples/2d_animation.py
@@ -1,6 +1,6 @@
-from blobmodel import Model, show_model, BlobShapeImpl, BlobShapeEnum
+from blobmodel import Model, show_model
 
-# here you can define your custom parameter distributions
+# Example of a simple two-dimensional realization and animation.
 
 bm = Model(
     Nx=100,
@@ -15,7 +15,7 @@ bm = Model(
     t_init=10,
 )
 
-# create data
+# Make a realization and store it in a DataSet object
 ds = bm.make_realization(speed_up=True, error=1e-2)
 # show animation and save as gif
 show_model(dataset=ds, interval=100, gif_name="2d_animation.gif", fps=10)
