### 2025-05-20 c467e47
Fix model stop on speed up and test
