+++ blobmodel/model.py
@@ -339,9 +339,9 @@ class Model:
         # ignores t_drain when calculating stop time
         stop = np.minimum(
             self._geometry.t.size,
-            start
-            + int(
-                (
+            int(
+                blob.t_init
+                + (
                     -blob.width_prop * np.log(error * np.sqrt(np.pi))
                     + self._geometry.Lx
                     - blob.pos_x
