### 2025-05-20 d4f65f8
Update paper/paper.md
Co-authored-by: Emily Bourne <louise.bourne@gmail.com>
