### 2025-05-20 0b0369c
Update paper/paper.md
Co-authored-by: Emily Bourne <louise.bourne@gmail.com>
