### 2025-05-20 dfd9fa2
Update paper/paper.md
Co-authored-by: Emily Bourne <louise.bourne@gmail.com>
