+++ paper/paper.bib
@@ -165,7 +165,7 @@
 
 @article{kociscak_modeling_2023,
 	author = {{Kočiščák, S.} and {Kvammen, A.} and {Mann, I.} and {Sørbye, S. H.} and {Theodorsen, A.} and {Zaslavsky, A.}},
-	title = {Modeling Solar Orbiter dust detection rates in the inner heliosphere as a Poisson process},
+	title = {Modeling Solar Orbiter dust detection rates in the inner heliosphere as a {Poisson} process},
 	DOI= "10.1051/0004-6361/202245165",
 	url= "https://doi.org/10.1051/0004-6361/202245165",
 	journal = {A&A},
