### 2025-05-26 gregordecristoforo/2d_propagating_blobs#118 @EmilyBourne
Display problem for long mathematical expression in docstring
On the following line:
https://github.com/uit-cosmo/blobmodel/blob/bc566ccb97421c3f03625f37c85fe04ed5427ec0/blobmodel/blobs.py#L17

The expression has become detached from its `math::` tag (probably due to the length).

This leads to display problems in the docs:

![Image](https://github.com/user-attachments/assets/54501ec9-86c7-47ad-92b7-bbe5208f683a)

There are a couple of other expressions in this docstring that are badly formatted.

This was noticed during the review of https://github.com/openjournals/joss-reviews/issues/8032 but I don't consider it blocking for the review. The API docs are there and it is trivial to find the raw code to read the docstring.
