### 2025-06-05 #8032 @fruzsinaagocs
@Sosnowsky, thanks for addressing the issues. I think the only remaining open one is https://github.com/uit-cosmo/blobmodel/issues/118.

@andrewgiuliani do you feel that your comments have been addressed?
