### 2025-06-16 #8032 @EmilyBourne
I am also waiting for a small edit (typo) to the paper:
https://github.com/uit-cosmo/blobmodel/pull/99#discussion_r2107574516
