### 2025-06-17 1762c19
Update paper/paper.md
Co-authored-by: Emily Bourne <louise.bourne@gmail.com>
