### 2025-06-23 #8032 @andrewgiuliani
would the authors be able to merge [this pull request](https://github.com/uit-cosmo/blobmodel/pull/99) so that editorialbot can compile the latest version of their paper write up? It's hard to read the tex source.
