### 2025-06-24 #8032 @EmilyBourne
@andrewgiuliani the branch indicated for the paper is the one in the PR so I'm not sure merging makes a difference. You just need to ask the bot to generate the pdf here
