### 2025-06-24 #8032 @fruzsinaagocs
Thank you both @EmilyBourne  and @andrewgiuliani  for the review and congratulations @Sosnowsky ! I'll generate a new proof and ask that you check that all references, author list, title, are as they should be so that we can proceed with acceptance.
