### 2025-07-07 #8032 @fruzsinaagocs
@Sosnowsky All looks good on my and as well. At this point could you:
