### 2025-07-08 #8032 @Sosnowsky
Awesome, I created a new release and it's now published in Zenodo (and Pypi, if relevant)
The new version tag is v1.1.1
DOI: 10.5281/zenodo.15836074
Let me know if there is anything else needed.
Thanks!
