### 2025-07-09 #8032 @fruzsinaagocs
@Sosnowsky just one more thing: can you please mark v1.1.1 as "Latest" among the releases?
