### 2025-07-10 #8032 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
